// src/app/types.ts
export interface LoginTexts {
  pageTitle: string;
  email?: string;
  password?: string;
  currentPassword?: string;
  newPassword?: string;
  newPasswordConfirm?: string;
  change?: string;
  guideTitle?: string;
  guide1?: string;
  guide2?: string;
  guide3?: string;
  guide4?: string;
  show?: string;
  hide?: string;
  showPassword?: string;
  hidePassword?: string;
  showNewPassword?: string;
  hideNewPassword?: string;
  showNewPasswordConfirm?: string;
  hideNewPasswordConfirm?: string;
  code?: string;
  sendCode?: string;
  verifying?: string;
  verify?: string;
  validationRequired?: string;
  validationPolicy?: string;
  validationMismatch?: string;
  success?: string;
  error?: string;
  sent?: string;
  leftTitle?: string;
  resultsTitle?: string;
  noneSelected?: string;
  resetButton?: string;
  allRegions?: string;
  itemsCount?: string;
  viewDetail?: string;
  selectAll?: string;
  deleteSelected?: string;
  listView?: string;
  galleryView?: string;
  searchButton?: string; 
  category?: string;
  region?: string;
  searchKeyword?: string;
  searchPlaceholder?: string;
  allCategories?: string;
  invalidEmail?: string;
  enterCode?: string;
  incorrectCode?: string;
  enterPassword?: string;
  forgotPassword?: string;
  login?: string;
  noAccount?: string;
  signup?: string;
  required?: string;
  proceed?: string;
  enterEmail?: string;
}